Reorganized tests to match code under test.
