+++ test/transaction/types_test.py
@@ -1,0 +1,41 @@
+# coding=utf-8
+from __future__ import absolute_import, division, print_function, \
+  unicode_literals
+
+from unittest import TestCase
+
+from six import binary_type
+
+from iota import TransactionHash
+
+
+class TransactionHashTestCase(TestCase):
+  def test_init_automatic_pad(self):
+    """
+    Transaction hashes are automatically padded to 81 trytes.
+    """
+    # noinspection SpellCheckingInspection
+    txn = TransactionHash(
+      b'JVMTDGDPDFYHMZPMWEKKANBQSLSDTIIHAYQUMZOK'
+      b'HXXXGJHJDQPOMDOMNRDKYCZRUFZROZDADTHZC'
+    )
+
+    # noinspection SpellCheckingInspection
+    self.assertEqual(
+      binary_type(txn),
+
+      # Note the extra 9's added to the end.
+      b'JVMTDGDPDFYHMZPMWEKKANBQSLSDTIIHAYQUMZOK'
+      b'HXXXGJHJDQPOMDOMNRDKYCZRUFZROZDADTHZC9999'
+    )
+
+  def test_init_error_too_long(self):
+    """
+    Attempting to create a transaction hash longer than 81 trytes.
+    """
+    with self.assertRaises(ValueError):
+      # noinspection SpellCheckingInspection
+      TransactionHash(
+        b'JVMTDGDPDFYHMZPMWEKKANBQSLSDTIIHAYQUMZOK'
+        b'HXXXGJHJDQPOMDOMNRDKYCZRUFZROZDADTHZC99999'
+      )
